Consolidating ApiManagement packages into the actual ApiManagement folders.  This would only be a problem for anyone not using a Windows git client.
If you happened to use Git for WSL, it ends up creating both the ApiManagement and APIManagement folders, like git was told to do.  Then, if you use that folder from Windows to load the solution, it gets confused, shows both folders, and only includes a subset of the files.