Github PR: https://github.com/Microsoft/BusinessPlatformApps/pull/983 had many comments that I addressed.
Two changes were discussed in a meeting: Credentials being specified through the appropriate mechanism (in both the integration tests and the actual MSFT->SG master key in the RetrieveSocialGistApiKey class).

The colors hardcoded in PopulateRedditUserDefinedEntities.cs were explained as well (in line with the commentary in the PR) and left as-is, but there's every change we will be removing them and doing them in another way entirely in the weeks to come.